requirements.txt
# QuantFlow — Combined Python Dependencies
# Install with: pip install -r requirements.txt

# ── Anthropic Claude AI ───────────────────────────────────────
anthropic

# ── Market Data ───────────────────────────────────────────────
yfinance

# ── Data Processing ───────────────────────────────────────────
pandas
numpy

# ── Charts ────────────────────────────────────────────────────
matplotlib
mplfinance

# ── Scheduling ────────────────────────────────────────────────
schedule

# ── Configuration ─────────────────────────────────────────────
python-dotenv

# ── Strategy Builder web app ──────────────────────────────────
# Used by Strategy_builder/app.py for the live backtest dashboard
fastapi
uvicorn[standard]
pydantic
